pyproject.toml
[tool.poetry]
name = "metadock"
version = "0.1.0"
description = ""
authors = ["David Sillman <david.sillman@intusurg.com>"]
readme = "README.md"
packages = [{ include = "metadock" }]

[tool.poetry.dependencies]
python = "^3.11"
jinja2 = "^3.1.2"
markdown = "^3.5.1"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[tool.poetry.scripts]
metadock = "metadock.cli:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
